gfs-browser/requirements.txt
flask
flask_accept
flask_swagger
flask_cors
pymongo
rdflib-jsonld
